Compiler version checking, flow
