+++ Sources/SassEmbedded/Compiler.swift
@@ -14,7 +14,7 @@ import Logging
 // Compiler -- interface, control state machine
 // CompilerChild -- Child process, NIO reads and writes
 // CompilerWork -- Sass stuff, protocol, job management
-// Compilation -- job state, many, managed by CompilerWork
+// CompilationRequest -- job state, many, managed by CompilerWork
 
 /// A Sass compiler interface that uses the Sass embedded protocol.
 ///
@@ -38,6 +38,8 @@ public final class Compiler {
     enum State {
         /// No child, new jobs wait on promise with new state.
         case initializing(EventLoopFuture<Void>)
+        /// Child up, checking it's working before accepting compilations.
+        case checking(CompilerChild, EventLoopFuture<Void>)
         /// Child is running and accepting compilation jobs.
         case running(CompilerChild)
         /// Child is broken.  Fail new jobs with the error.  Reinit permitted.
@@ -50,8 +52,15 @@ public final class Compiler {
 
         var child: CompilerChild? {
             switch self {
-            case .running(let c), .quiescing(let c, _): return c
+            case .checking(let c, _), .running(let c), .quiescing(let c, _): return c
             case .initializing(_), .broken(_), .shutdown: return nil
+            }
+        }
+
+        var future: EventLoopFuture<Void>? {
+            switch self {
+            case .initializing(let f), .checking(_, let f), .quiescing(_, let f): return f
+            case .running, .broken, .shutdown: return nil
             }
         }
 
@@ -61,8 +70,16 @@ public final class Compiler {
             return future
         }
 
+        mutating func inittingToChecking(_ child: CompilerChild) {
+            self = .checking(child, future!)
+        }
+
+        mutating func checkingToRunning() {
+            self = .running(child!)
+        }
+
         mutating func toQuiescing(_ future: EventLoopFuture<Void>) -> EventLoopFuture<Void> {
-            self = .quiescing(child!, future) // ! -> must be running to start quiesce
+            self = .quiescing(child!, future)
             return future
         }
     }
@@ -218,7 +235,7 @@ public final class Compiler {
             switch state {
             case .broken, .shutdown:
                 return eventLoop.makeSucceededFuture(nil)
-            case .running(let child), .quiescing(let child, _):
+            case .checking(let child, _), .running(let child), .quiescing(let child, _):
                 return eventLoop.makeSucceededFuture(child.processIdentifier)
             case .initializing(let future):
                 return future.flatMap {
@@ -289,7 +306,7 @@ public final class Compiler {
                     m.syntax = .init(syntax)
                     url.flatMap { m.url = $0.absoluteString }
                     asyncImporter.flatMap {
-                        m.importer = .init($0, id: Compilation.baseImporterID)
+                        m.importer = .init($0, id: CompilationRequest.baseImporterID)
                     }
                 }),
                 outputStyle: outputStyle,
@@ -331,15 +348,13 @@ public final class Compiler {
             // jobs submitted after/during shutdown: fail them.
             work.cancelAllPending(with: LifecycleError("Compiler has been shutdown, not accepting further work"))
 
-        case .initializing:
+        case .initializing, .checking:
             // jobs submitted while [re]starting the compiler: wait.
             break
 
         case .running(let child):
             // goodpath
-            work.startAllPending().forEach {
-                child.send(message: $0)
-            }
+            child.send(messages: work.startAllPending())
         }
     }
 
@@ -350,7 +365,7 @@ public final class Compiler {
     ///
     /// When this future completes the system is either in running or broken.
     private func startCompiler() -> EventLoopFuture<Void> {
-        precondition(!work.hasActiveCompilations)
+        precondition(!work.hasActiveRequests)
         startCount += 1
         return initThread.runIfActive(eventLoop: eventLoop) { [self] in
             try CompilerChild(eventLoop: eventLoop,
@@ -359,9 +374,13 @@ public final class Compiler {
                               errorHandler: { [unowned self] in self.handle(error: $0) })
         }.flatMap { child in
             child.addChannelHandlers()
-        }.map { child in
-            self.debug("Compiler is started")
-            self.state = .running(child)
+        }.flatMap { child -> EventLoopFuture<Versions> in
+            self.debug("Compiler is started, starting healthcheck")
+            self.state.inittingToChecking(child)
+            return child.sendVersionRequest()
+        }.map { versions in
+            // version check here
+            self.state.checkingToRunning()
             self.kickPendingCompilations()
         }.flatMapErrorThrowing { error in
             self.debug("Can't start the compiler at all: \(error)")
@@ -377,7 +396,7 @@ public final class Compiler {
     /// 1. Write transport errors, reported by a promise from `CompilerChild.send(message:to:)`
     /// 2. Read transport errors, reported by the channel handler from `CompilerChild.errorCaught(...)`
     /// 3. Protocol errors reported by the Sass compiler, from `CompilerWork.receieveGlobal(message:)`
-    /// 4. Protocol errors detected by us, from `Compilation.receive(message)`.
+    /// 4. Protocol errors detected by us, from `CompilationRequest.receive(message)`.
     /// 5. User-injected restarts, from `reinit()`.
     /// 6. Timeouts, from `CompilerWork`'s reset API.
     ///
@@ -389,6 +408,13 @@ public final class Compiler {
 
         switch state {
         case .initializing(let future):
+            return future
+
+        case .checking(let child, let future):
+            // Timeout or something while checking the version, kick the process
+            // and let the init process handle the error.
+            debug("Error while checking compiler, stopping it")
+            child.stopAndCancelWork(with: error)
             return future
 
         case .running(let child):
@@ -423,8 +449,8 @@ public final class Compiler {
         eventLoop.preconditionInEventLoop()
 
         switch state {
-        case .initializing(let initFuture):
-            debug("Shutdown during restart, deferring")
+        case .initializing(let initFuture), .checking(_, let initFuture):
+            debug("Shutdown during startup, deferring")
             // Nasty corner - wait for the init to resolve and then
             // try again!
             return initFuture.flatMap { _ in
@@ -523,6 +549,27 @@ final class CompilerChild: ChannelInboundHandler {
             self.errorHandler(ProtocolError("Write to Sass compiler failed: \(error)."))
             return self.eventLoop.makeFailedFuture(error)
         }
+    }
+
+    func send(messages: [Sass_EmbeddedProtocol_InboundMessage]) {
+        messages.forEach { send(message: $0) }
+    }
+
+    func sendVersionRequest() -> EventLoopFuture<Versions> {
+        let (future, _) = work.startVersionRequest()
+// TODO: When the compiler actually supports Version, let it speak.
+//       Until then, fake up the response.
+        eventLoop.scheduleTask(in: .milliseconds(100)) {
+            var rmsg = Sass_EmbeddedProtocol_OutboundMessage()
+            rmsg.versionResponse = .with { ver in
+                ver.compilerVersion = "1"
+                ver.protocolVersion = "2"
+                ver.implementationName = "Dummy"
+                ver.implementationVersion = "13"
+            }
+            self.receive(message: rmsg)
+        }
+        return future
     }
 
     /// Called from the pipeline handler with a new message
